feat(ir): widen IntConst payload u64→u128, drop Result from build_int_const
NodeKind::IntConst now stores u128 instead of u64, enabling full-width
constant folding for U128 values.  build_int_const becomes infallible
(no Result, no ErrorKind::IntConstWidthExceedsU64), accepts impl Into<u128>,
and masks the value to the declared NodeOutputType width.
build_uint64_const is removed; all ~365 call sites are updated to drop
the trailing ? / .unwrap().

eval_int_binary / eval_int_cmp in opt are widened to u128 arithmetic;
pattern Bindings::Int and Match::get_int / get_int_const widen to u128;
int_const_with_fn closure return type widens to Result<u128>.  KnownBits
keeps u64 internal masks and casts safely, guarded by the existing
U128/U256 early-out.  Integer literal suffixes are added across all test
and bench files to satisfy the new impl Into<u128> bound.

+++ crates/analyzer/tests/common/mod.rs
@@ -178,7 +178,8 @@ pub fn has_kind<F: Fn(&NodeKind) -> bool>(g: &ir::BuiltFunctionGraph, pred: F) -
 }
 
 pub fn has_constant(g: &ir::BuiltFunctionGraph, value: u64) -> bool {
-    has_kind(g, |k| matches!(k, NodeKind::IntConst(c) if *c == value))
+    // IntConst stores u128; compare against the u64 value widened to u128.
+    has_kind(g, |k| matches!(k, NodeKind::IntConst(c) if *c == u128::from(value)))
 }
 
 // ── per_arch_test! macro ─────────────────────────────────────────────────────
